Use file types returned by read_dir rather than requerying path
This has two effects:

1) The file types returned by read_dir don't follow symbolic links. So
whereas Path::is_dir would return true if the path was a symbolic
link to a directory, the file_type of the directory entry would be a
symbolic link, so would return false to is_dir. This means that this
change prevents the tool from following symbolic links. Following
symbolic links can lead to processing of parts of the filesystem that
the user didn't intend and can also lead to effectively running forever
if a symbolic link leads to a parent directory (as is the case if you
have a rustc build directory lying around).

2) By avoiding requerying whether the directory entry is a directory, we
avoid unnecessary syscalls. This speeds up runtime on one of my directories
from 827ms to 415ms.