Add --depth argument to limit the depth of the search for `target/` directories
I'm not super familiar with crossbeam so maybe the implementation is not idiomatic.

Tbh results are not that impressive, with `--depth 2`, I get 11s in my usecase, will without it, I have 14s.

However the difference is more noticeable for machine with few cores, as setting `--threads 1`, I see an improvement from 22s to 11s by enabling `--depth 2`.